PICO: maths performance - use pico_float and pico_double from SDK (#14663)
* Maths performance: PICO more complete use of pico_float and pico_double from SDK

* Wrap floating point math functions and some double arithmetic with efficient pico-sdk versions.
* Also replace sqrt with sqrtf in scheduler (under USE_LATE_TASK_STATISTICS).

* PICO: maths performance - fix build

* PICO RP2350.mk tidy up and add wrap for __ctzdi2

* PICO makefile don't repeat -Wl,--wrap=

+++ src/main/scheduler/scheduler.c
@@ -601,7 +601,7 @@ FAST_CODE void scheduler(void)
                 // 10ths % of tasks late in last second
                 DEBUG_SET(DEBUG_TIMING_ACCURACY, 4, lateTaskPercentage);
 
-                float gyroCyclesStdDev = sqrt(devSquared/gyroCyclesCount);
+                float gyroCyclesStdDev = sqrtf(devSquared/gyroCyclesCount);
                 int32_t gyroCyclesStdDev100thus = clockCyclesTo100thMicros((int32_t)gyroCyclesStdDev);
                 DEBUG_SET(DEBUG_TIMING_ACCURACY, 7, gyroCyclesStdDev100thus);
                 DEBUG_SET(DEBUG_SCHEDULER_DETERMINISM, 7, gyroCyclesStdDev100thus);
